UI updated to only load uploaded content instead re-uploading every media file

+++ app/views/media_uploader.py
@@ -96,7 +96,8 @@ def remove_media(path):
         print("path: ", path)
         print("username: ", username)
         database.remove(db, Media, path)
-        os.remove(path)
+        if (os.path.exists(path)):
+            os.remove(path)
         # medias = database.retrieve_from_join(db, User, Media, username)
         # file_paths = [database.retrieve(Media, media_id=media.media_id).path for media in medias]
